embed header in block; move getters to header struct (#421)

+++ testutil/timestamp.go
@@ -9,11 +9,11 @@ package testutil
 import "github.com/facebookgo/clock"
 
 // TimestampNow get now timestamp from new clock
-func TimestampNow() uint64 {
+func TimestampNow() int64 {
 	return TimestampNowFromClock(clock.New())
 }
 
 // TimestampNowFromClock get now timestamp from specific clock
-func TimestampNowFromClock(c clock.Clock) uint64 {
-	return uint64(c.Now().Unix())
+func TimestampNowFromClock(c clock.Clock) int64 {
+	return c.Now().Unix()
 }
